+++ .env
@@ -1,0 +1,1 @@
+SECRET_KEY="8udexni@$w&i)p5we4jz!4xpesil92lq1-)ng4pirxxn!elob0"
