+++ .gitignore
@@ -2,3 +2,6 @@
 *.faiss
 vectordb/full_patent_metadata.jsonl
 
+/migrations/*.py
+/migrations/*.pyc
+!**/migrations/init.py
