+++ pass_project/accounts/urls.py
@@ -5,5 +5,5 @@ app_name = 'accounts'
 urlpatterns = [
     path('login/', views.login_view, name='login'),
     path('mypage/', views.mypage_view, name='mypage'),
-    path('mypage/repassword/', views.repassword_view, name='repassword'),
+    path('repassword/', views.repassword_view, name='repassword'),
 ]
