+++ pass_project/core/static/core/js/main.js
@@ -13,12 +13,12 @@ function goToAIassistpage() {
 function goToAIqapage() {
   console.log("✅ goToAIqapage() 실행됨");
 
-  window.location.href = "http://127.0.0.1:8000/1";  // 원하는 주소로 수정  
+  window.location.href = "http://127.0.0.1:8000/2";  // 원하는 주소로 수정  
 }
 
 function goToLoginpage() {
   console.log("✅ goToLoginpage() 실행됨");
 
-  window.location.href = "http://127.0.0.1:8000/3";  // 원하는 주소로 수정
+  window.location.href = "http://127.0.0.1:8000/accounts/login/";  // 원하는 주소로 수정
 }
