+++ pass_project/pass_project/urls.py
@@ -21,10 +21,12 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('core.urls')),           # 메인 페이지
+
     path('accounts/', include('accounts.urls')), # 로그인, 마이페이지
-    path('user_admin/', include('user_admin.urls')),    # 관리자 페이지
-    path('assist/', include('assist.urls')),  # AI 기능
+
+    path('', include('core.urls')),           # 메인 페이지
+    # path('user_admin/', include('user_admin.urls')),    # 관리자 페이지
+    # path('assist/', include('assist.urls')),  # AI 기능
 ]
 
 if settings.DEBUG:
